code/requirements.txt
certifi>=2024.0.0
charset-normalizer>=3.0.0
idna>=3.0
joblib>=1.3.0
networkx>=3.0
numpy>=1.24.0,<2.0  # matplotlib requires numpy<2
pandas>=2.0.0
python-dateutil>=2.8.0
pytz>=2024.0
requests>=2.30.0
scikit-learn>=1.3.0
scipy>=1.10.0
six>=1.16.0
threadpoolctl>=3.0.0
tqdm>=4.65.0
tzdata>=2024.0
urllib3>=2.0.0
python-louvain==0.16  # For community detection/modularity
python-dotenv>=1.0.0  # For .env file support
matplotlib>=3.7.0     # For plotting
seaborn>=0.12.0       # For statistical visualizations